Update booker.php
This is the recommended entry for the online Paypal booking operation mode of booker.html and the one that should be advertised to users. In the event that booker.html changes, the php callto booker.php will ensure the latest version is always used rather than out-of-date-cache.  The efficiency-loss can be disregarded in the circumstances